cli: split commands/workspace.rs into sub modules
